requirements.txt
configparser~=5.1.0
Flask~=2.0.2
mysql-connector-python~=8.0.27